--- eBPF_Supermarket/Stack_Analyser/libbpf/框架使用方法.md
@@ -1,31 +1,0 @@
-# eBPF侧
-
-1. 在 `bpf/` 文件夹中添加自定义eBPF程序代码，命名方式为 `子功能名.bpf.c`
-
-2. 在代码中使用 `DECLARE_MAPS(type)` 宏声明公用的map并以 `type` 指定 count 值的类型
-
-3. 使用 `KERNEL_STACK` 和 `USER_STACK` 来获取内核和用户调用栈id，调用栈会自动保存在 stack_trace 中
-
-# 用户侧
-
-1. 在 `include/stack_analyzer.h` 中的 `MOD` 枚举类型定义中新增一个功能模块的标识，将该名称添加到`MOD_NUM`前面即可
-
-2. 在 `stack_analyzer.cc` 中创建一个 `bpf_loader` 的子类，在其中重写以下函数：
-
-    1. 自定义eBPF程序加载函数 `int load(void)` 加载自定义eBPF程序
-
-    2. 自定义eBPF程序挂载函数 `int attach(void)` 设置eBPF程序相应参数并挂载到相应跟踪点
-
-    3. 遍历 count map ，会使用自定义解析函数 `uint64_t data_value(void *count)` 将count解析为单值value并产生以value值升序的有序的 sortedData(pid, usid, ksid, count) 列表
-
-    4. 若不需要实时展示火焰图，则将sortedData列表输出，每条输出会使用自定义注解函数 `std::string data_str(uint64_t value)` 为value值添加对应的单位和名称
-
-    5. 自定义eBPF程序卸载函数 `void detach(void)`，将eBPF程序与跟踪点分离，停止采集
-
-    6. 自定义eBPF程序清除函数 `void remove(void)`，将eBPF程序清除
-
-3. 在 main 函数中添加新增子命令和对应参数解析语句
-
-# 编译侧
-
-在 `Makefile` 文件中的 `APPS` 变量中添加 新增的子功能名
